# Biconnectivity module using visitor functions