--- a/src/Biconnectivity/Biconnectivity.jl
+++ b/src/Biconnectivity/Biconnectivity.jl
@@ -1,0 +1,13 @@
+module Biconnectivity
+
+using LightGraphs
+using SimpleTraits
+using LightGraphs.SimpleGraphs: SimpleEdge
+using LightGraphs.Traversals: traverse_graph!, DepthFirst, TraversalState
+import LightGraphs.Traversals: previsitfn!, newvisitfn!, revisitfn!
+
+include("articulation.jl")
+include("bridge.jl")
+include("biconnect.jl")
+
+end # module