--- a/src/Biconnectivity/Biconnectivity.jl
+++ b/src/Biconnectivity/Biconnectivity.jl
@@ -5,6 +5,7 @@ using SimpleTraits
 using LightGraphs.SimpleGraphs: SimpleEdge
 using LightGraphs.Traversals: traverse_graph!, DepthFirst, TraversalState
 import LightGraphs.Traversals: previsitfn!, newvisitfn!, revisitfn!
+using LightGraphs.Traversals: VSUCCESS
 
 include("articulation.jl")
 include("bridge.jl")