**Benchmarks** 

There is a slowdown of about 20%, but I'm trying to fix that.

New -
```julia
julia> g = loadsnap(:facebook_combined)
{4039, 88234} undirected simple Int64 graph

julia> @btime LightGraphs.Biconnectivity.articulation(g)
  1.441 ms (28 allocations: 192.73 KiB)

julia> @btime LightGraphs.Biconnectivity.bridges(g)
  1.463 ms (29 allocations: 195.95 KiB)

julia> @btime LightGraphs.Biconnectivity.biconnected_components(g)
  4.205 ms (248 allocations: 4.59 MiB)
```
Old -
```julia

julia> g1 = loadsnap(:facebook_combined)
{4039, 88234} undirected simple Int64 graph

julia> @btime articulation(g)
  1.135 ms (22 allocations: 160.92 KiB)

julia> @btime bridges(g)
  1.170 ms (22 allocations: 163.98 KiB)

julia> @btime biconnected_components(g)
  3.782 ms (231 allocations: 4.46 MiB)
```