@sbromberger 2020-06-05
As discussed - we should make sure the edges are ordered.