@sbromberger 2020-06-05
Also - are those benchmarks you posted using the incorrect old version of biconnected_components, or the fixed version?