@abhinavmehndiratta 2020-06-05
> Also - are those benchmarks you posted using the incorrect old version of biconnected_components, or the fixed version?

The fixed version. Also, after merging the new visitor return codes, this PR will need some changes though (changing `return true` to `return VSUCCESS`).