@abhinavmehndiratta 2020-06-05
> As discussed - we should make sure the edges are ordered.

Okay, I'll do that. But still need to figure out how to make this fast.