@sbromberger 2020-06-05
Yeah. I'm concerned that this is slower than the fixed version, which was itself slower than the incorrect version.